Unload gem specs before loading the desired sub-check
The behavior of `Gem.activate_bin_path` changed starting in Ruby 2.4.5. Instead
of only loading this gem, it now loads all its deepndencies as well. This
can result in dependency conflicts when, say, this check has loaded sensu-plugin
4.0 and the sub-check hasn't yet been updated to support > 1.2.

This also bumps the sensu-plugin dep to support the recent 4.x release.